Make the modal a component
