+++ src/components/ApplyTicket/ApplyTicket.css
@@ -41,18 +41,3 @@
 .ticket-inclusion-item {
     width: 30%;
 }
-
-.form-check-input:checked {
-    background-color: #c83e38 !important;
-    border-color: #c83e38 !important;
-}
-
-.ticket-cancel-btn {
-    border: 1px solid #888888;
-}
-
-.ticket-cancel-btn:hover,
-.ticket-cancel-btn:focus {
-    border-color: #c83e38;
-    color: #c83e38;
-}
